+++ python_ta/__init__.py
@@ -29,7 +29,6 @@ except AttributeError:
 
 
 import logging
-import tokenize
 import webbrowser
 from typing import IO, Any, Optional, Union