+++ python_ta/checkers/forbidden_io_function_checker.py
@@ -1,9 +1,12 @@
 """Checker for use of I/O functions.
 """
 
-from astroid import nodes
+from re import sub
+from typing import Union
+
+from astroid import BoundMethod, FunctionDef, nodes
 from pylint.checkers import BaseChecker
-from pylint.checkers.utils import only_required_for_messages
+from pylint.checkers.utils import only_required_for_messages, safe_infer
 from pylint.lint import PyLinter
 
 from python_ta.utils import _is_in_main
@@ -21,8 +24,8 @@ class IOFunctionChecker(BaseChecker):
         "E9998": (
             "Used input/output function %s",
             "forbidden-IO-function",
-            'Used when you use the I/O functions "print", "open" or "input". These '
-            "functions should not be used except where specified by your instructor.",
+            'Used when you use the I/O functions "print", "open" or "input" or other config-specified forbidden '
+            "functions. These functions should not be used except where specified by your instructor.",
         )
     }
     options = (
@@ -32,7 +35,7 @@ class IOFunctionChecker(BaseChecker):
                 "default": FORBIDDEN_BUILTIN,
                 "type": "csv",
                 "metavar": "<builtin function names>",
-                "help": "List of built-in function names that should not be used.",
+                "help": "List of I/O function names and method qualified names that should not be used.",
             },
         ),
         (
@@ -48,31 +51,34 @@ class IOFunctionChecker(BaseChecker):
 
     @only_required_for_messages("forbidden-IO-function")
     def visit_call(self, node: nodes.Call) -> None:
-        if isinstance(node.func, nodes.Name):
-            name = node.func.name
-            # ignore the name if it's not a builtin (i.e. not defined in the
-            # locals nor globals scope)
-            if not (name in node.frame() or name in node.root()):
-                scope = node.scope()
-                scope_parent = scope.parent
+        if (name := self._resolve_qualname(node)) in self.linter.config.forbidden_io_functions:
+            scope = node.scope()
+            scope_parent = scope.parent
 
-                if (
-                    isinstance(scope_parent, nodes.ClassDef)
-                    and isinstance(scope, nodes.FunctionDef)
-                    and (scope_parent.name + "." + scope.name) not in self.linter.config.allowed_io
-                ):
-                    if name in self.linter.config.forbidden_io_functions:
-                        self.add_message("forbidden-IO-function", node=node, args=name)
-                elif (
-                    isinstance(scope_parent, nodes.Module)
-                    and isinstance(scope, nodes.FunctionDef)
-                    and scope.name not in self.linter.config.allowed_io
-                ):
-                    if name in self.linter.config.forbidden_io_functions:
-                        self.add_message("forbidden-IO-function", node=node, args=name)
-                elif isinstance(scope, nodes.Module) and not _is_in_main(node):
-                    if name in self.linter.config.forbidden_io_functions:
-                        self.add_message("forbidden-IO-function", node=node, args=name)
+            if (
+                isinstance(scope_parent, nodes.ClassDef)
+                and isinstance(scope, nodes.FunctionDef)
+                and (scope_parent.name + "." + scope.name) not in self.linter.config.allowed_io
+            ):
+                self.add_message("forbidden-IO-function", node=node, args=name)
+            elif (
+                isinstance(scope_parent, nodes.Module)
+                and isinstance(scope, nodes.FunctionDef)
+                and scope.name not in self.linter.config.allowed_io
+            ):
+                self.add_message("forbidden-IO-function", node=node, args=name)
+            elif isinstance(scope, nodes.Module) and not _is_in_main(node):
+                self.add_message("forbidden-IO-function", node=node, args=name)
+
+    @staticmethod
+    def _resolve_qualname(node: nodes.Call) -> Union[str, None]:
+        """Resolves the qualified name for function and method calls"""
+        if (inferred_definition := safe_infer(node.func)) is not None:
+            if isinstance(inferred_definition, (BoundMethod, FunctionDef)):
+                return sub(r"^[^.]*\.", "", inferred_definition.qname())
+        if isinstance(node.func, nodes.Name):
+            return node.func.name
+        return None
 
 
 def register(linter: PyLinter) -> None: