+++ CHANGELOG.md
@@ -12,10 +12,15 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Integrated Watchdog to enable automatic re-checking of Python files when changes are detected.
 - Added `autoformat-options` configuration option to let users specify command-line arguments to the Black formatting tool
 - Update `check_all` and `check_error` functions to let users pass in `typing.IO` objects to the `output` argument
+- Update the `forbidden-io-function-checker` to check functions from imported modules as well as methods (according to their qualified name)
+- Update the `forbidden-io-function-checker` to flag aliases of forbidden functions
+- Update how error messages are overridden such that section headers are no longer required within the config file
 
 ### 💫 New checkers
 
 ### 🐛 Bug fixes
+
+- `check_contracts` no longer makes methods immediately enforce Representation Invariant checks when setting attributes of instances with the same type (one `Node` modifies another `Node` instance) and only checks RIs for these instances after the method returns.
 
 ### 🔧 Internal changes